Make sure to always set the image field
This fixes a regression where newly generated notes in an editing session would end up with empty cards due to a missing image field. The reason for this lay in the assumption that edited cards would use the same image, if not changed via the change image button. Of course this could not possibly apply to newly generated cards which didn't have an image field set in the first place.