Set up add-on configuration and templates at profile initialization
Up to this point the add-on config/templates would only be set up
upon user interaction (e.g. creating new I/O notes or changing options).
This commit will allows us to update the note type before a user has
the chance to interact with potentially outdated notes/cards.

(of importance for instance with the switch to Anki 2.1, where
updates to the card templates are needed for the cards to render
correctly in the Reviewer)