Only load original image after mask loading has completed
Fixes #45. Most likely also fixes #15.

Even with a liberal fadein timer set, the image flashing problem
would still persist on slower systems and/or systems that have
performance issues with the new web engine (macOS?).
Postponing image visibility until after the mask is loaded
seems like the only surefire way to prevent instances of the answers
flashing before the actual answer reveal.

We still use a small aFade timer to make transitions between question
and answer where the images have not been cached smoother.

Because this change comes with a lot of increased complexity to the
basic method of operation of the add-on, and because it has the potential
to completely disrupt card reviewing, it will require extensive
testing across all platforms (especially the mobile clients and
AnkiWeb).