Confirmation popup before closing the occlusion editor
Fixes #72.
Note that this confirmation popup is shown in some cases where I don't
think it's necessary. If the user adds occlusions, clicks on one of the
buttons to create the cards and then closes the editor dialog they are
prompted for confirmation.
The only way I can think of to prevent this is to reset the svg-edit
undo stack when the user clicks on the button, but I'm not sure if
that's acceptable.