Disable locale support for now
Due to how quickly Anki's architecture and add-on APIs are evolving, it does not seem feasible to pursue locale support for now (including e.g. the time it would cost to crowd-source and validate translations, modify the build system to support translations, etc.).

These changes are as non-destructive as possible, allowing us to drop locale support back in at any point in the future.